Extracting internals module
