+++ mulang.cabal
@@ -95,6 +95,7 @@ library
     Language.Mulang.Edl.Lexer
     Language.Mulang.Edl.Parser
     Language.Mulang.Interpreter
+    Language.Mulang.Interpreter.Internals
     Language.Mulang.Interpreter.Runner
 
   build-depends:
